+++ src/a7.components.js
@@ -213,6 +213,9 @@ User.prototype.getMemento = function(){
 function View( props ){
 	this.renderer = a7.model.get("a7").ui.renderer;
 	this.type = 'View';
+	this.timeout;
+	this.timer;
+	this.element;
 	this.props = props;
 	this.state = {};
 	this.mustRender = false;
@@ -233,6 +236,11 @@ View.prototype = {
 		}.bind( this ));
 
 		this.on( "rendered", function(){
+			// set the timeout
+			if( this.timer !== undefined ){
+				clearTimeout( this.timer );
+			}
+			this.timer = setTimeout( this.checkRenderStatus.bind( this ), a7.model.get( "a7" ).ui.timeout );
 			this.mustRender = false;
 			this.onRendered();
 		}.bind( this ));
@@ -261,17 +269,17 @@ View.prototype = {
 	},
 	render: function(){
 		a7.log.info( 'render: ' + this.props.id );
-		if( this.props.element === undefined || this.props.element === null ){
-			this.props.element = document.querySelector( this.props.selector );
+		if( this.element === undefined || this.element === null ){
+			this.element = document.querySelector( this.props.selector );
 		}
-		if( !this.props.element ) throw( "You must define a selector for the view." );
-		this.props.element.innerHTML = ( typeof this.template == "function" ? this.template() : this.template );
+		if( !this.element ) throw( "You must define a selector for the view." );
+		this.element.innerHTML = ( typeof this.template == "function" ? this.template() : this.template );
 
 		var eventArr = [];
 		a7.ui.getEvents().forEach( function( eve ){
 			eventArr.push("[data-on" + eve + "]");
 		});
-		var eles = this.props.element.querySelectorAll( eventArr.toString() );
+		var eles = this.element.querySelectorAll( eventArr.toString() );
 
 		eles.forEach( function( sel ){
 			for( var ix=0; ix < sel.attributes.length; ix++ ){
@@ -288,9 +296,16 @@ View.prototype = {
 	onRendered: function(){
 		if( this.props !== undefined && this.props.children !== undefined && this.props.children !== null ){
 			for( var child in this.props.children ){
-				this.props.children[ child ].props.element = document.querySelector( this.props.children[ child ].props.selector );
+				this.props.children[ child ].element = document.querySelector( this.props.children[ child ].props.selector );
 				this.props.children[ child ].render();
 			}
+		}
+	},
+	checkRenderStatus: function(){
+		if( document.querySelector( this.props.selector ) === null ){
+			a7.ui.unregister( this.id );
+		}else{
+			this.timer = setTimeout( this.checkRenderStatus.bind( this ), a7.model.get( "a7" ).ui.timeout );
 		}
 	}
 };